Created a room_manager.py file
