+++ BE/routers/rooms.py
@@ -1,15 +1,15 @@
 import asyncio
-import json
 from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
 from typing import Dict, List
 import uuid
 from pydantic import BaseModel
 from routers.data_manager import load_data, save_data
+from routers.room_manager import load_active_rooms, save_active_rooms, discard_room_id
 router = APIRouter()
 
 rooms_data: Dict[str, Dict[str, List[Dict[str, str]]]] = load_data("rooms_data.json")
-ACTIVE_ROOMS_FILE = "active_rooms.json"
+
 
 active_rooms: List[str] = load_data("active_rooms.json")
 room_timers: Dict[str, asyncio.TimerHandle] = {}
@@ -18,31 +18,6 @@ class joinRoomParams(BaseModel):
     user_id: str
     user_name: str
 
-def load_active_rooms():
-    try:
-        with open(ACTIVE_ROOMS_FILE, "r") as file:
-            return json.load(file)
-    except FileNotFoundError:
-        return []   
-    
-def save_active_rooms(active_rooms_list):
-    with open(ACTIVE_ROOMS_FILE, "w") as file:
-        json.dump(active_rooms_list, file)
-
-def discard_room_id(room_id):
-    global room_timers, rooms_data, active_rooms
-
-    if room_id in room_timers:
-        room_timers[room_id].cancel()
-        del room_timers[room_id]
-
-    if room_id in rooms_data:
-        del rooms_data[room_id]
-        save_data("rooms_data.json", rooms_data)
-    
-    if room_id in active_rooms:
-        active_rooms.remove(room_id)
-        save_active_rooms(active_rooms)
 
 @router.post("/create_room", response_model=Dict[str, str])
 async def create_room():
